Spotify Popularity Predictor and Explorer
DJ Khaled boldly claimed to always know when a song will be a hit. We decided to further investigate this by                     asking three key questions: Are there certain characteristics for hit songs, what are the largest influencers on a                  song’s success, and whether old songs even predict the popularity of new songs? Predicting how popular a                 song will be is no easy task.